Github action for running backend tests
