+++ backend/test/factories/users.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :user do
     email { Faker::Internet.email }
-    password { 'password' }
-    password_confirmation { 'password' }
+    password { 'Valid1!!' }
+    password_confirmation { 'Valid1!!' }
   end
 end
